Search Bar Display Logic (#730)
* hides search bar when no attributes are searchable

* adds tests for search bar display logic
* extracts capture_view_locals method to ControllerHelpers

* no need for page argument to show_search_bar? method

* streamlines show_search_bar? method
