+++ app/views/administrate/application/index.html.erb
@@ -17,6 +17,8 @@ It renders the `_table` partial to display details about the resources.
   By default, these resources are passed to the table partial to be displayed.
 - `search_term`:
   A string containing the term the user has searched for, if any.
+- `show_search_bar`:
+  A boolean that determines if the search bar should be shown.
 
 [1]: http://www.rubydoc.info/gems/administrate/Administrate/Page/Collection
 %>
@@ -26,7 +28,9 @@ It renders the `_table` partial to display details about the resources.
 <% end %>
 
 <% content_for(:search) do %>
-  <%= render "search", search_term: search_term %>
+  <% if show_search_bar %>
+    <%= render "search", search_term: search_term %>
+  <% end %>
 <% end %>
 
 <header class="header">
