+++ spec/controllers/admin/customers_controller_spec.rb
@@ -22,6 +22,13 @@ describe Admin::CustomersController, type: :controller do
 
       expect(locals[:page]).to be_instance_of(Administrate::Page::Collection)
     end
+
+    it "shows the search bar" do
+      customer = create(:customer)
+
+      locals = capture_view_locals { get :index }
+      expect(locals[:show_search_bar]).to be_truthy
+    end
   end
 
   describe "GET show" do
@@ -162,18 +169,5 @@ describe Admin::CustomersController, type: :controller do
 
       expect(response).to redirect_to(admin_customers_url)
     end
-  end
-
-  def capture_view_locals
-    allow(@controller).to receive(:render)
-    yield
-
-    locals = nil
-    expect(@controller).to have_received(:render).at_least(1).times do |*args|
-      args.each do |arg|
-        locals ||= arg.try(:fetch, :locals, nil)
-      end
-    end
-    locals
   end
 end
